fix a rare bug which is caused by precision lost. In wyTileSetInfoGetRect method, we calculate tile coordinate based on image width and tile width, but image width is cutted when loading, we need keep the precision so change wyUtils::loadXXX parameter from size_t* to float*. Also lua binding is updated

+++ jni/WiEngine/impl/platform/macosx/wyUtils_macosx.h
@@ -75,7 +75,7 @@ public:
 	 * @return 适合于创建OpenGL贴图的数据, 如果\c sizeOnly是true, 则返回NULL, 调用者需要负责释放返回的数据
 	 * \endif
 	 */
-	static char* loadNSImage(NSImage* image, size_t* w, size_t* h, bool sizeOnly, float scaleX, float scaleY);
+	static char* loadNSImage(NSImage* image, float* w, float* h, bool sizeOnly, float scaleX, float scaleY);
 	
 	/**
 	 * \if English
