+++ jni/WiEngine/impl/utils/wyUtils.cpp
@@ -967,7 +967,7 @@ const char* wyUtils::decodeObfuscatedData(const char* data, size_t length, size_
 	}
 }
 
-char* wyUtils::loadBMP(FILE* f, size_t* w, size_t* h, bool sizeOnly, float scaleX, float scaleY) {
+char* wyUtils::loadBMP(FILE* f, float* w, float* h, bool sizeOnly, float scaleX, float scaleY) {
 	// load raw data
 	size_t length;
 	char* raw = loadRaw(f, &length);
@@ -980,7 +980,7 @@ char* wyUtils::loadBMP(FILE* f, size_t* w, size_t* h, bool sizeOnly, float scale
 	return data;
 }
 
-char* wyUtils::loadBMP(int resId, size_t* w, size_t* h, bool sizeOnly) {
+char* wyUtils::loadBMP(int resId, float* w, float* h, bool sizeOnly) {
 	// load raw data
 	size_t length;
 	float scale;
@@ -994,7 +994,7 @@ char* wyUtils::loadBMP(int resId, size_t* w, size_t* h, bool sizeOnly) {
 	return data;
 }
 
-char* wyUtils::loadBMP(const char* path, bool isFile, size_t* w, size_t* h, bool sizeOnly, float scaleX, float scaleY) {
+char* wyUtils::loadBMP(const char* path, bool isFile, float* w, float* h, bool sizeOnly, float scaleX, float scaleY) {
 	// load raw data
 	size_t length;
 	char* raw = loadRaw(path, isFile, &length);
@@ -1007,7 +1007,7 @@ char* wyUtils::loadBMP(const char* path, bool isFile, size_t* w, size_t* h, bool
 	return data;
 }
 
-char* wyUtils::loadPNG(FILE* f, size_t* w, size_t* h, bool sizeOnly, float scaleX, float scaleY) {
+char* wyUtils::loadPNG(FILE* f, float* w, float* h, bool sizeOnly, float scaleX, float scaleY) {
 	// load raw data
 	size_t length;
 	char* raw = loadRaw(f, &length);
@@ -1020,7 +1020,7 @@ char* wyUtils::loadPNG(FILE* f, size_t* w, size_t* h, bool sizeOnly, float scale
 	return data;
 }
 
-char* wyUtils::loadPNG(int resId, size_t* w, size_t* h, bool sizeOnly) {
+char* wyUtils::loadPNG(int resId, float* w, float* h, bool sizeOnly) {
 	// load raw data
 	size_t length;
 	float scale;
@@ -1034,7 +1034,7 @@ char* wyUtils::loadPNG(int resId, size_t* w, size_t* h, bool sizeOnly) {
 	return data;
 }
 
-char* wyUtils::loadPNG(const char* path, bool isFile, size_t* w, size_t* h, bool sizeOnly, float scaleX, float scaleY) {
+char* wyUtils::loadPNG(const char* path, bool isFile, float* w, float* h, bool sizeOnly, float scaleX, float scaleY) {
 	// load raw data
 	size_t length;
 	char* raw = loadRaw(path, isFile, &length);
@@ -1047,7 +1047,7 @@ char* wyUtils::loadPNG(const char* path, bool isFile, size_t* w, size_t* h, bool
 	return data;
 }
 
-char* wyUtils::loadJPG(FILE* f, size_t* w, size_t* h, bool sizeOnly, float scaleX, float scaleY) {
+char* wyUtils::loadJPG(FILE* f, float* w, float* h, bool sizeOnly, float scaleX, float scaleY) {
 	// load raw data
 	size_t length;
 	char* raw = loadRaw(f, &length);
@@ -1060,7 +1060,7 @@ char* wyUtils::loadJPG(FILE* f, size_t* w, size_t* h, bool sizeOnly, float scale
 	return data;
 }
 
-char* wyUtils::loadJPG(int resId, size_t* w, size_t* h, bool sizeOnly) {
+char* wyUtils::loadJPG(int resId, float* w, float* h, bool sizeOnly) {
 	// load raw data
 	size_t length;
 	float scale;
@@ -1074,7 +1074,7 @@ char* wyUtils::loadJPG(int resId, size_t* w, size_t* h, bool sizeOnly) {
 	return data;
 }
 
-char* wyUtils::loadJPG(const char* path, bool isFile, size_t* w, size_t* h, bool sizeOnly, float scaleX, float scaleY) {
+char* wyUtils::loadJPG(const char* path, bool isFile, float* w, float* h, bool sizeOnly, float scaleX, float scaleY) {
 	// load raw data
 	size_t length;
 	char* raw = loadRaw(path, isFile, &length);
@@ -1131,7 +1131,7 @@ char* wyUtils::loadRaw(FILE* f, size_t* outLen, bool noDecode) {
 	return data;
 }
 
-bool wyUtils::getPVRSize(const char* data, size_t length, size_t* w, size_t* h, float scale) {
+bool wyUtils::getPVRSize(const char* data, size_t length, float* w, float* h, float scale) {
 	// create buffer
 	wyPVRHeader* header = NULL;
 	if ((header = (wyPVRHeader*)wyMalloc(sizeof(wyPVRHeader))) == NULL) {
@@ -1157,8 +1157,8 @@ bool wyUtils::getPVRSize(const char* data, size_t length, size_t* w, size_t* h, 
 	// check format if supports scaling
 //	wyPVRFormat format = (wyPVRFormat)(header->flags & PVR_TEXTURE_FLAG_TYPE_MASK);
 //	if(canScalePVR(format)) {
-//		*w = *w * scale + 0.5f;
-//		*h = *h * scale + 0.5f;
+//		*w = *w * scale;
+//		*h = *h * scale;
 //	}
 	
 	// free header
@@ -1167,7 +1167,7 @@ bool wyUtils::getPVRSize(const char* data, size_t length, size_t* w, size_t* h, 
 	return true;
 }
 
-bool wyUtils::getPVRSize(FILE* f, size_t* w, size_t* h, float scale) {
+bool wyUtils::getPVRSize(FILE* f, float* w, float* h, float scale) {
 	// load raw data
 	size_t length;
 	char* raw = loadRaw(f, &length);
@@ -1181,7 +1181,7 @@ bool wyUtils::getPVRSize(FILE* f, size_t* w, size_t* h, float scale) {
 	return ret;
 }
 
-bool wyUtils::getPVRSize(int resId, size_t* w, size_t* h, float* outScale) {
+bool wyUtils::getPVRSize(int resId, float* w, float* h, float* outScale) {
 	// load raw data
 	size_t length;
 	char* raw = loadRaw(resId, &length, outScale);
@@ -1194,7 +1194,7 @@ bool wyUtils::getPVRSize(int resId, size_t* w, size_t* h, float* outScale) {
 	return ret;
 }
 
-bool wyUtils::getPVRSize(const char* path, bool isFile, size_t* w, size_t* h, float scale) {
+bool wyUtils::getPVRSize(const char* path, bool isFile, float* w, float* h, float scale) {
 	// load raw data
 	size_t length;
 	char* raw = loadRaw(path, isFile, &length);
